Add files via upload
Within the purple curve, we can see a representation of a 3D loop within the DNA. The annotation (green) shows us which genes are present within this section of DNA, which we can check expression of this gene within the RNA activity section (red). The ATAC sections (blue) show us the open chromatin regions, which we can cross-reference with the non-coding open chromatin regions to see which open regions are actually coding. We can also see in blue the exon regions, which line up with the RNA activity section to denote transcription.